requirements.txt
fastapi==0.111.0
nba_api>=1.4.1
uvicorn[standard]==0.29.0
httpx==0.27.0
scipy==1.11.4
numpy==1.26.4
anthropic>=0.40.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
